feat: add tests for pokemonApi
